Add messages
Messages have one sender and may have one or more receivers. On creation,
a status is created for the sender and each of the receivers. When the last
of these statuses gets deleted, their associated message gets deleted, too.
Currently, the status only keeps track of the read status for each player.