Make buildings, researches, and units easily configurable
An initializer is used to setup buildings etc. via calling
LaFamiglia.buildings.add etc. When there are code changes to the
module LaFamiglia, the webserver has to be restarted because
otherwise the configuration done in the initializer is lost
due to the constant LaFamiglia being reloaded.

Rename lafamiglia.rb to la_famiglia.rb to make autoloading work

Move usage of LaFamiglia.config.unit_speed to Unit#speed